Moving the ReadSocketResponse in a adhoc PayloadReader class
