+++ src/csharp/Microsoft.Spark.Worker.UnitTest/DaemonWorkerTests.cs
@@ -49,7 +49,7 @@ namespace Microsoft.Spark.Worker.UnitTest
             var rowsReceived = new List<object[]>();
 
             WriteTestData(clientSocket);
-            ReadSocketResponse(
+            PayloadReader.Read(
                 clientSocket.InputStream, 
                 rowsReceived, 
                 ref timingDataReceived, 
@@ -95,50 +95,6 @@ namespace Microsoft.Spark.Worker.UnitTest
             SerDe.Write(outputStream, (int)SpecialLengths.END_OF_DATA_SECTION);
             SerDe.Write(outputStream, (int)SpecialLengths.END_OF_STREAM);
             outputStream.Flush();
-        }
-
-        private static void ReadSocketResponse(
-            Stream inputStream, 
-            List<object[]> rowsReceived, 
-            ref bool timingDataReceived,
-            ref bool exceptionThrown)
-        {
-            while (true)
-            {
-                var length = SerDe.ReadInt32(inputStream);
-                if (length > 0)
-                {
-                    byte[] pickledBytes = SerDe.ReadBytes(inputStream, length);
-                    var unpickler = new Unpickler();
-
-                    var rows = unpickler.loads(pickledBytes) as ArrayList;
-                    foreach (object row in rows)
-                    {
-                        rowsReceived.Add((object[]) row);
-                    }
-                }
-                else if (length == (int)SpecialLengths.TIMING_DATA)
-                {
-                    long bootTime = SerDe.ReadInt64(inputStream);
-                    long initTime = SerDe.ReadInt64(inputStream);
-                    long finishTime = SerDe.ReadInt64(inputStream);
-                    long memoryBytesSpilled = SerDe.ReadInt64(inputStream);
-                    long diskBytesSpilled = SerDe.ReadInt64(inputStream);
-                    timingDataReceived = true;
-                }
-                else if (length == (int)SpecialLengths.PYTHON_EXCEPTION_THROWN)
-                {
-                    SerDe.ReadString(inputStream);
-                    exceptionThrown = true;
-                    break;
-                }
-                else if (length == (int)SpecialLengths.END_OF_DATA_SECTION)
-                {
-                    int numAccumulatorUpdates = SerDe.ReadInt32(inputStream);
-                    SerDe.ReadInt32(inputStream);
-                    break;
-                }
-            }
         }
     }
 }
